services/vector_cache/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
fastapi==0.129.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
numpy==2.4.2
portalocker==3.2.0
prometheus_client==0.24.1
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
qdrant-client==1.16.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
